+++ examples/example-project/src/user.clj
@@ -1,9 +1,5 @@
 (ns user
-  (:require [scicloj.clay.v2.api :as clay]
-            [scicloj.kindly-default.v1.api :as kindly-default]))
-
-;; Initialize Kindly's [default](https://github.com/scicloj/kindly-default).
-(kindly-default/setup!)
+  (:require [scicloj.clay.v2.api :as clay]))
 
 ;; Start Clay.
 (clay/start!)