+++ src/scicloj/clay/v2/tool/scittle.clj
@@ -1,8 +1,5 @@
 (ns scicloj.clay.v2.tool.scittle
-  (:require [scicloj.kindly.v3.api :as kindly]
-            [scicloj.kindly.v3.kind :as kind]
-            [scicloj.kindly.v3.kindness :as kindness]
-            [scicloj.clay.v2.walk]
+  (:require [scicloj.clay.v2.walk]
             [scicloj.clay.v2.tool :as tool]
             [hiccup.core :as hiccup]
             [hiccup.page :as page]